Split goto_function_templatet into a separate file
