+++ src/goto-programs/goto_functions_template.h
@@ -14,6 +14,7 @@ Date: June 2003
 #ifndef CPROVER_GOTO_PROGRAMS_GOTO_FUNCTIONS_TEMPLATE_H
 #define CPROVER_GOTO_PROGRAMS_GOTO_FUNCTIONS_TEMPLATE_H
 
+#include "goto_function_template.h"
 #include <ostream>
 #include <cassert>
 
@@ -21,79 +22,6 @@ Date: June 2003
 #include <util/symbol.h>
 #include <util/cprover_prefix.h>
 
-template <class bodyT>
-class goto_function_templatet
-{
-public:
-  bodyT body;
-  code_typet type;
-
-  typedef std::vector<irep_idt> parameter_identifierst;
-  parameter_identifierst parameter_identifiers;
-
-  bool body_available() const
-  {
-    return !body.instructions.empty();
-  }
-
-  bool is_inlined() const
-  {
-    return type.get_bool(ID_C_inlined);
-  }
-
-  bool is_hidden() const
-  {
-    return type.get_bool(ID_C_hide);
-  }
-
-  void make_hidden()
-  {
-    type.set(ID_C_hide, true);
-  }
-
-  goto_function_templatet()
-  {
-  }
-
-  void clear()
-  {
-    body.clear();
-    type.clear();
-    parameter_identifiers.clear();
-  }
-
-  void swap(goto_function_templatet &other)
-  {
-    body.swap(other.body);
-    type.swap(other.type);
-    parameter_identifiers.swap(other.parameter_identifiers);
-  }
-
-  void copy_from(const goto_function_templatet &other)
-  {
-    body.copy_from(other.body);
-    type=other.type;
-    parameter_identifiers=other.parameter_identifiers;
-  }
-
-  goto_function_templatet(const goto_function_templatet &)=delete;
-  goto_function_templatet &operator=(const goto_function_templatet &)=delete;
-
-  goto_function_templatet(goto_function_templatet &&other):
-    body(std::move(other.body)),
-    type(std::move(other.type)),
-    parameter_identifiers(std::move(other.parameter_identifiers))
-  {
-  }
-
-  goto_function_templatet &operator=(goto_function_templatet &&other)
-  {
-    body=std::move(other.body);
-    type=std::move(other.type);
-    parameter_identifiers=std::move(other.parameter_identifiers);
-    return *this;
-  }
-};
 
 template <class bodyT>
 class goto_functions_templatet
